requirements.txt
# ──────────────────────────────────────────────────────────────────────────────
# Optional viewer: Obsidian (https://obsidian.md)
#   Free desktop app (Windows / Mac / Linux) for browsing and reviewing the
#   generated Markdown outputs, including spaced-repetition flashcard review.
#   NOT a Python package — install separately from obsidian.md.
#   Setup: open the output/ directory as an Obsidian vault, enable plugins,
#   then open summaries/flashcards.md → card icon → "Review Flashcards".
# ──────────────────────────────────────────────────────────────────────────────

annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.86.0
anyio==4.13.0
certifi==2026.2.25
click==8.3.1
colorama==0.4.6
distro==1.9.0
docstring_parser==0.17.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
lxml==6.0.2
markdown-it-py==4.0.0
mdurl==0.1.2
pillow==12.1.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
PyMuPDF==1.27.2.2
python-dotenv==1.2.2
python-pptx==1.0.2
rich==14.3.3
shellingham==1.5.4
sniffio==1.3.1
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
xlsxwriter==3.2.9
